Revert "Avoid any warning with recents clang *and* gcc"
This reverts commit 8b9e6ad494e2cda37e4965acc4afd6679f4d8d30.

+++ blosc/bitshuffle-avx2.h
@@ -20,16 +20,16 @@ extern "C" {
 /**
   AVX2-accelerated bitshuffle routine.
 */
-BLOSC_NO_EXPORT int32_t
-    bshuf_trans_bit_elem_avx2(void* in, void* out, const int32_t size,
-                              const int32_t elem_size, void* tmp_buf);
+BLOSC_NO_EXPORT int64_t
+    bshuf_trans_bit_elem_avx2(void* in, void* out, const size_t size,
+                              const size_t elem_size, void* tmp_buf);
 
 /**
   AVX2-accelerated bitunshuffle routine.
 */
-BLOSC_NO_EXPORT int32_t
-    bshuf_untrans_bit_elem_avx2(void* in, void* out, const int32_t size,
-                                const int32_t elem_size, void* tmp_buf);
+BLOSC_NO_EXPORT int64_t
+    bshuf_untrans_bit_elem_avx2(void* in, void* out, const size_t size,
+                                const size_t elem_size, void* tmp_buf);
 
 #ifdef __cplusplus
 }
